+++ armadillo4rtest/src/10_sparse_matrices.h
@@ -10,18 +10,18 @@
   return as_dgCMatrix(A);
 }
 
-// [[cpp4r::register]] SEXP sum_matrices_(SEXP x) {
-//   // Convert from dgCMatrix to SpMat
-//   SpMat<double> A = as_SpMat(x);
+[[cpp4r::register]] SEXP test_sparse_to_dgCMatrix(SEXP x) {
+  // Accepts any sparse Matrix object (CsparseMatrix, RsparseMatrix,
+  // TsparseMatrix, symmetric, triangular, ...) and round-trips it as a
+  // dgCMatrix.
+  SpMat<double> A = as_SpMat(x);
+  return as_dgCMatrix(A);
+}
 
-//   // Create a matrix B with a diagonal of random numbers
-//   SpMat<double> B(A.n_rows, A.n_cols);
-//   for (uword i = 0; i < A.n_rows; ++i) {
-//     B(i, i) = randu<double>();
-//   }
-
-//   A += B;  // Add the two matrices
-
-//   // Convert back to dgCMatrix and return
-//   return as_dgCMatrix(A);
-// }
+[[cpp4r::register]] SEXP test_sparse_dims(SEXP x) {
+  SpMat<double> A = as_SpMat(x);
+  writable::integers out(2);
+  out[0] = static_cast<int>(A.n_rows);
+  out[1] = static_cast<int>(A.n_cols);
+  return out;
+}